Optimize, prefer Bourne Shell, and make safer
I thought perhaps with `bash` having so many features, maybe it's
possible for running with `sh` (or compatibility mode in bash) to be
safer, due to there being fewer bells and whistles, so less about which
to concern oneself.

Since `grep` takes all of its options as traditional UNIX-style
flags, I have AWK (running in POSIX and sandbox mode) only sending
arguments through to `grep` which begin with at least one hyphen (`-`).
I've tried to circumvent this, but had no joy, thankfully.

I believe the code is cleaner without the redundant function juggling
and convoluted logic. No longer do you need to glance up and down the
script to find out what a single `if` statement does.

If I've forgotten or broken something, or missed a compatibility
issue, please let me know so that I can (hopefully) fix it.

By the way, this is a brilliant idea for a script! I may wind up
utilizing the code in this commit. As always, I appreciate the
inspiration.