Optimize via `printf` and remove `timestamp`
The `msg()` function is now more concise. The `date` command is great
for `sh` usage or when needing more advanced time and date features,
but `printf` itself is very powerful.

See `help printf` and `man printf` for information on both the built-in
and CoreUtils (for me, the latter is: `/usr/bin/printf`) versions.