Condense and optimize code
You might also be interested in this:

  https://github.com/terminalforlife/BashConfig/blob/master/.shplugs/Make_ScreenCast

Same concept, but gets around audio-video sync issues, and also applies
a noise filter based on the noise profile you gather outside of the
function (instructions in the script). Needs `sox` first, though.

Admittedly, it's not entirely finished yet, but it's functional, IIRC.