Simplify script via exhaustive `find` command
Okay, I recommend digging into the `find` manual page if you wish to
dig deep into the usage here.

In a nut-shell, this is a _lot_ faster and more efficient, thanks to
the use of `\+` reusing the command in the `-exec` section, meaning a
bazillion processes aren't launched.

The parentheses are used for grouping, as with shell's subshell `()`, -
but more akin to its standard grouping, `{}`. These are very important
to the workings of this `find` command.

Using the `-xdev` flag keeps `find` from visiting other filesystems.

Where you see `jonas:jonas`, it might be better to use your UID and
GID, respectively.